requirements.txt
python-dotenv
pandas
openpyxl
pyyaml
aiohttp
openai>=1.3.7
pyrogram
tgcrypto
aiosqlite
requests
httpx
pytz
schedule
python-telegram-bot
pyttsx3
langdetect
cryptography
tabulate
# 自动安装的依赖（完整部署系统）
pydantic-settings>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0
redis>=5.0.0
hiredis>=2.2.0
tiktoken>=0.5.0
aiofiles>=23.2.0
prometheus-client>=0.19.0

# Auto-installed packages
aiohappyeyeballs
aiosignal
annotated-types
anyio
attrs
certifi
cffi
charset-normalizer
distro
et_xmlfile
frozenlist
h11
httpcore
idna
jiter
multidict
numpy
pip
propcache
pyaes
pycparser
pydantic
pydantic_core
PySocks
python-dateutil
setuptools
six
sniffio
tqdm
typing_extensions
typing-inspection
tzdata
urllib3
wheel
yarl
